requirements.txt
aiogram==3.25.0
APScheduler==3.10.4
aiosqlite==0.20.0
httpx==0.28.1
feedparser==6.0.11
python-dotenv==1.0.1
PyYAML==6.0.2
beautifulsoup4==4.12.3
Telethon==1.41.2
